backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
python-multipart
requests
httpx
ollama